Restructure mathematical functions.
