+++ gqcp/include/Mathematical/AbstractFunction/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(gqcp
+    PRIVATE
+        LinearCombination.hpp
+        ScalarFunction.hpp
+)
