+++ gqcp/include/Mathematical/AbstractFunction/ScalarFunction.hpp
