+++ gqcp/include/Mathematical/CMakeLists.txt
@@ -1,10 +1,9 @@
 target_sources(gqcp
     PRIVATE
         CartesianDirection.hpp
-        LinearCombination.hpp
-        ScalarFunction.hpp
 )
 
+add_subdirectory(AbstractFunction)
 add_subdirectory(Algorithm)
 add_subdirectory(Grid)
 add_subdirectory(Optimization)
