+++ gqcp/include/Operator/SecondQuantized/USQOneElectronOperator.hpp
@@ -21,8 +21,8 @@
 #include "Basis/SpinorBasis/JacobiRotationParameters.hpp"
 #include "Basis/SpinorBasis/Spin.hpp"
 #include "Basis/TransformationMatrix.hpp"
+#include "Mathematical/AbstractFunction/ScalarFunction.hpp"
 #include "Mathematical/Representation/QCMatrix.hpp"
-#include "Mathematical/ScalarFunction.hpp"
 #include "Processing/RDM/OneRDM.hpp"
 #include "Processing/RDM/TwoRDM.hpp"
 #include "Utilities/type_traits.hpp"
