+++ gqcp/include/gqcp.hpp
@@ -51,6 +51,8 @@
 #include "Basis/SpinorBasis/USpinorBasis.hpp"
 #include "Basis/TransformationMatrix.hpp"
 #include "Basis/transform.hpp"
+#include "Mathematical/AbstractFunction/LinearCombination.hpp"
+#include "Mathematical/AbstractFunction/ScalarFunction.hpp"
 #include "Mathematical/Algorithm/Algorithm.hpp"
 #include "Mathematical/Algorithm/CompoundConvergenceCriterion.hpp"
 #include "Mathematical/Algorithm/ConvergenceCriterion.hpp"
@@ -61,7 +63,6 @@
 #include "Mathematical/CartesianDirection.hpp"
 #include "Mathematical/Grid/CubicGrid.hpp"
 #include "Mathematical/Grid/Field.hpp"
-#include "Mathematical/LinearCombination.hpp"
 #include "Mathematical/Optimization/Accelerator/ConstantDamper.hpp"
 #include "Mathematical/Optimization/Accelerator/DIIS.hpp"
 #include "Mathematical/Optimization/ConsecutiveIteratesNormConvergence.hpp"
@@ -101,7 +102,6 @@
 #include "Mathematical/Representation/SquareMatrix.hpp"
 #include "Mathematical/Representation/SquareRankFourTensor.hpp"
 #include "Mathematical/Representation/Tensor.hpp"
-#include "Mathematical/ScalarFunction.hpp"
 #include "Molecule/Molecule.hpp"
 #include "Molecule/NuclearFramework.hpp"
 #include "Molecule/Nucleus.hpp"
