+++ gqcp/tests/Mathematical/AbstractFunction/ScalarFunction_test.cpp
@@ -20,7 +20,7 @@
 #include <boost/test/unit_test.hpp>
 
 #include "Basis/ScalarBasis/CartesianGTO.hpp"
-#include "Mathematical/ScalarFunction.hpp"
+#include "Mathematical/AbstractFunction/ScalarFunction.hpp"
 
 
 BOOST_AUTO_TEST_CASE(ScalarFunctionProduct) {
