+++ gqcp/tests/Mathematical/CMakeLists.txt
@@ -1,7 +1,4 @@
-list(APPEND test_target_sources
-    ${CMAKE_CURRENT_SOURCE_DIR}/ScalarFunction_test.cpp
-)
-
+add_subdirectory(AbstractFunction)
 add_subdirectory(Grid)
 add_subdirectory(Optimization)
 add_subdirectory(Representation)
